tms320c5400: Fixed standard library selection.
This selects the standard library for all tms320c5400 toolsets.
Autolinking is used for version 4 and later.  Version 3 and earlier has
a very simplistic approach and uses rts.lib for all options.

This also removes quotes from the paths for version 3 and earlier.  This
doesn't seem to be a problem for version 4 and later at this point.